test(rn): share browser local query and subscription scenarios
